third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEOffsetElement-dom-dx-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'dx' attribute of the SVGFEOffsetElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS offsetElement.getAttribute('dx') is "50"
PASS offsetElement.getAttribute('dx') is "0"
PASS successfullyParsed is true

TEST COMPLETE
